The scraper accepts jobs comprised of a job title, a list of skills, and aliases for the skills and job title. An example alias would be "node.js" as for the skill "node". The scrape will record the presence of a skill in a post if the skill or any of its aliases are found.
Previously the scraper settings form provided a single input field for the submission of aliases. Input was automatically formatted into a csv. This took up excessive space on the space and yielded an unclear interface. This branch refactors the code to include popup forms for editing skill aliases. There is also substantial underlying refactoring of state. This yielded a design that is in the same vein as redux.

This code should be refactored further substantially.