Over the last two weeks I (barely) stitched together the architecture of the project. I should have made a larger number of smaller commits,but better late than never. Here is a list of the big changes.
1) RabbitMQ was added.
Why? Assuming multiple users are using the app, there is a need for functionality to control when jobs are scheduled. The dream scenario would be that a comprehensive schedule always be available (perhaps displayed in a user's frontend application), that jobs can be moved earlier and later in the schedule efficiently at will, that the schedule be repeated continously, etc.

To simplify things, the following scheme is being built using RabbitMQ:
Each user has a list of "jobs". A job is a collection of related job titles along with skills and aliases for those skills. For example:
job1: { titles: [software engineer, coder], skills: { python: python2, python3,    spark: apache-spark, spark2.4 } }.
When a user adds a new job, it is sent to a Nodejs backend that generates a unique key for the job, and then stores all the job info in a MongoDb database. Then Node places sends the key to a RabbitMQ queue. Meanwhile (discussed further below), there are python threads extracting keys from the front of the queue. When a thread extracts the key, it uses the key to look up the job info in Mongo and then scrapes data according to that info. For example, the thread might determine that there have been 5 new jobs posted with a title of "software engineer" or "coder" with at least one of the skills "python" and "spark" (could be that theyre listed with an alias0 since the last scraping. The thread collects the locations, date posted, and other info about the jobs and then inserts them into Mongo. Finally, the thread sends a message to Node that it finished the job. The message contains the job key of the completed job. Users can set jobs to active or inactive; active jobs are those that should have additional data scraped regularly. This active/inactive status is stored in Mongo. So, when Node receives a job key from a python thread, it uses the key to look up in Mongo if the job is active. If it is, Node inserts the key back into the queue, and so the cycle repeats.

On shutdown, the current order of tasks is cached. On startup, it is retrieved and reinserted to the queue

Downsides/reasons why this scheme might be changed later: 1) You cannot change the order of keys in the queue at will. 2) If a user flips a job to inactive, it will run one more time before it is actually halted.

Notes: RabbitMQ supports routing of messages to different queues as well as priority queues. This allows future flexibility if need be.

2) Celery
The python scraper threads are implemented using celery. Basically Celery is a book keeping process that spins up worker threads whenever there are keys in the Rabbit queue. Celery is fantastic at scaling for tasks like scraping. (This will not be done for moral reasons), but Celery can quickly and efficiently spin up thousands of scraper threads.

3) Tor
While celery can quickly spin up thousands of scraper threads, they will quickly be blocked by Indeed. To deal with this, Tor functionality is being added. Essentially celery has Tor running next to it with a local Socks proxy. Therefore the threads can shoot their requests through this proxy rather than sending them directly to Indeed. Then, if Indeed blocks, the local Tor process can simply be restarted, which will provide a new IP address to scrape from.

Right now, celery processes are paired with their own tor processes, and there are two celery processes. This might need to be rearranged.

Note 1: Using tor slows down the requests. But without, there's a danger of having an IP blocked permanently. Also, I did a bottleneck analysis and it was silly how much pausing sufficiently to avoid this danger constrained overall performance.

Note 2: I have zero intention to abuse this. Scraping will be set to reasonable rates so as to not weigh down the Indeed servers. Further, I'm looking into means of slowing or ceasing scraping when traffic is high. I do want protection though that the scraping can be dialed down and the app continued if I accidentally scrape too quickly.

4) Docker
To simplify dependencies and development and to learn about an awesome technology, I've dockerized everything. This would make it easy to scale by spinning up more services (probably the node service to maintain the queue or more celery/tor processes).

5) Core scraping code
The core scraping code has been written

Next up: a huge amount of cleaning up, actually filling out the front end react code, writing the node code (write now it just forwards requests to rabbit), converting the scraping code to work with celery, implementing my sketches of the Mongo schema, writing the backend code to retrieve data from Mongo and format it for d3 charts on the front end.